pyproject.toml
[tool.poetry]
name = "softs-implementation"
version = "0.1.0"
description = "Implementation of SOFTS: Efficient Multivariate Time Series Forecasting with Series-Core Fusion"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"

packages = [{ include = "softs_implementation", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
torch = ">=1.10.0"
pandas = ">=1.3.0"
numpy = ">=1.21.0"
scikit-learn = ">=1.0.0"
PyYAML = ">=6.0"
tqdm = ">=4.62.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
ipykernel = ">=6.0.0"
flake8 = ">=6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
